feat: add --with-locations to noir-inspector print-acir
Move the ACIR opcode annotation builder from noirc_driver into a new
noirc_artifacts::annotations module so it can work directly on compiled
artifacts, and use it in `noir-inspector print-acir --with-locations`.
Since artifacts embed both the debug symbols and the source file map,
the ACIR of an existing artifact can be annotated with the originating
Noir locations and snippets without the source tree or a recompile.
Works for both program and contract artifacts.
