+++ tooling/inspector/Cargo.toml
@@ -24,6 +24,7 @@ serde_json.workspace = true
 color-eyre.workspace = true
 const_format.workspace = true
 acir.workspace = true
+fm.workspace = true
 noirc_artifacts.workspace = true
 noirc_artifacts_info.workspace = true
 noir_artifact_cli.workspace = true
